move mocked_file_manager_factory to conftest.py
