+++ jobbergate-api/jobbergate_api/tests/apps/job_scripts/test_job_script_files.py
@@ -3,10 +3,8 @@ Test job-script files.
 """
 
 from pathlib import Path
-from unittest.mock import patch
 
 import pytest
-from file_storehouse.engine import EngineLocal
 
 from jobbergate_api.apps.job_scripts.job_script_files import (
     JOBSCRIPTS_MAIN_FILE_FOLDER,
@@ -14,22 +12,6 @@ from jobbergate_api.apps.job_scripts.job_script_files import (
     JOBSCRIPTS_WORK_DIR,
     JobScriptFiles,
 )
-
-
-@pytest.fixture(scope="function")
-def mocked_file_manager_factory(tmp_path):
-    """
-    Fixture to replace the default file engine (EngineS3) by a local one.
-
-    In this way, all objects are stored in a temporary directory ``tmp_path``,
-    that is yield for reference.
-    """
-
-    def local_engine_factory(*, work_directory: Path, **kwargs):
-        return EngineLocal(base_path=tmp_path / work_directory)
-
-    with patch("jobbergate_api.s3_manager.engine_factory", wraps=local_engine_factory):
-        yield tmp_path
 
 
 class TestJobScriptFiles:
